% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.surveyresponses.r
\name{as.data.frame.surveyresponses}
\alias{as.data.frame.surveyresponses}
\title{as.data.frame.surveyresponses}
\usage{
\method{as.data.frame}{surveyresponses}(survey)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A data frame with survey responses
}
\description{
Extracts data from the survey responses data set and formats it as a data frame for analysis
}

