% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyresponses.r
\name{clean_sm_labels}
\alias{clean_sm_labels}
\title{Get sensible variable names for an \code{sm_survey} object.}
\usage{
clean_sm_labels(survey)
}
\arguments{
\item{survey}{an \code{sm_survey} object, as retrieved by \code{surveylist()}}
}
\value{
a tibble with question_id and subquestion_id and metadata columns related to main and sub-questions and order of appearance in the survey.
}
\description{
Called internally by \code{survey_responses()}, but can also be used interactively.
}
