% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_survey_by_id.R
\name{find_survey_by_id}
\alias{find_survey_by_id}
\title{find_survey_by_id}
\usage{
find_survey_by_id(id, s)
}
\arguments{
\item{id}{A numeric survey monkey survey_id of the survey to be returned}

\item{s}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A single of object of class {sm_response}
}
\description{
Get responses for a SurveyMonkey survey
}
