% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_survey_by_name.R
\name{find_survey_by_name}
\alias{find_survey_by_name}
\title{Fetch a survey by name.}
\usage{
find_survey_by_name(s_title, survey_list)
}
\arguments{
\item{s_title}{the exact name of the survey to retrieve.}

\item{survey_list}{the survey list to retrieve from}
}
\value{
the desired sm_survey object.
}
\description{
Supply the name of your survey and a survey_list and get back the \code{sm_survey} object.
}
\examples{
# not run:
# s <- survey_list(per_page = 200)
# my_survey <- find_survey_by_name("survey for testing", s)
}
