\name{getresponses}
\alias{getresponses}
\alias{print.sm_response}
\title{Get responses}
\description{Get responses for a survey}
\usage{
getresponses(respondents, survey,
             api_key = getOption('sm_api_key'),
             oauth_token = getOption('sm_oauth_token'))
}
\arguments{
\item{respondents}{A vector containing respondent ID numbers, possibly returned by \code{\link{respondentlist}}.}
\item{survey}{A survey ID number, possibly returned by \code{\link{surveylist}}.}
\item{api_key}{Your API key. By default, retrieved from \code{getOption('sm_api_key')}.}
\item{oauth_token}{Your OAuth 2.0 token, as generated by \code{\link{smlogin}}. By default, retrieved from \code{getOption('sm_oauth_token')}.}
%  \item{...}{Other arguments. Currently ignored.}
}
\details{Retrieves response data for requested respondents to a specified survey collector.}
\value{A list containing one or more objects of class \code{sm_response}.}
\references{
\url{https://developer.surveymonkey.com/mashery/get_responses}
}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{
smlogin()
s <- surveylist()
r <- respondentlist(s[1])
# getresponses(r)
}
}
%\keyword{}
