% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_questions.R
\name{survey_choices}
\alias{survey_choices}
\title{survey_choices}
\usage{
survey_choices(survey)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A data frame with one row per choice
}
\description{
Creates a data frame from the survey choices
}
