% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_details.r
\name{survey_details}
\alias{survey_details}
\title{survey_details}
\usage{
survey_details(survey, question_details = TRUE,
  oauth_token = getOption("sm_oauth_token"), ...)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}

\item{oauth_token}{Your OAuth 2.0 token, as generated by \code{smlogin}. By default, retrieved from \code{getOption('sm_oauth_token')}.}

\item{question_detail}{A logical value to indicate whether to include details on questions and answer choices.  By default = TRUE.}
}
\value{
A list of objects of class \code{sm_survey}.
}
\description{
Get detailed information about a survey
}
\references{
SurveyMonkey API V3 at \url{https://developer.surveymonkey.com/api/v3/#surveys}
}
