% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_list.r
\name{survey_list}
\alias{survey_list}
\title{survey_list}
\usage{
survey_list(page = NULL, per_page = NULL, sort_by = NULL,
  sort_order = NULL, start_modified_at = NULL,
  end_modified_at = NULL, title = NULL, include = NULL,
  oauth_token = getOption("sm_oauth_token"), ...)
}
\arguments{
\item{page}{Integer number to select which page of resources to return. By default is 1.}

\item{per_page}{Integer number to set the number of surveys to return per page. By default, is 50 surveys per page.}

\item{sort_by}{String used to sort returned survey list: 'title', 'date_modified', or 'num_responses'. By default, date_modified.}

\item{sort_order}{String used to set the sort order for returned surveys: 'ASC' or 'DESC'. By default, DESC.}

\item{start_modified_at}{Date string used to select surveys last modified after this date. By default is NULL.}

\item{end_modified_at}{Date string used to select surveys modified before this date.  By default is NULL.}

\item{title}{String used to select survey by survey title.  By default is NULL.}

\item{include}{Comma separated strings used to filter survey list: 'shared_with', 'shared_by', or 'owned' (useful for teams) or to specify additional fields to return per survey: 'response_count', 'date_created', 'date_modified', 'language', 'question_count', 'analyze_url', 'preview'.  By default is NULL.}

\item{oauth_token}{Your OAuth 2.0 token, as generated by \code{smlogin}. By default, retrieved from \code{getOption('sm_oauth_token')}.}
}
\value{
A list of objects of class \code{sm_survey}.
}
\description{
Get the list of the user's surveys. This function calls the SurveyMonkey API using the current oauth token and returns a list of surveys filtered by the parameters entered.
}
\references{
SurveyMonkey API V3 at \url{https://developer.surveymonkey.com/api/v3/#surveys}

Rmonkey at \url{https://github.com/sfirke/Rmonkey}
}
