% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyquestiondf.R
\name{surveyquestions}
\alias{surveyquestions}
\title{surveyquestions}
\usage{
surveyquestions(survey)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A data frame with one row per question/subquestion/answer choice
}
\description{
Creates a data frame from the survey questions and answers
}

