% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyresponses.r
\name{surveyresponses}
\alias{surveyresponses}
\title{surveyresponses}
\usage{
surveyresponses(survey)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A data frame with survey responses
}
\description{
Extracts data from the survey responses data set and formats it as a data frame for analysis
}

