\name{userdetails}
\alias{userdetails}
\title{Get User Details}
\description{Once logged in, get basic details about the user account.}
\usage{
userdetails(api_key = getOption('sm_api_key'),
            oauth_token = getOption('sm_oauth_token'), ...)
}
\arguments{
\item{api_key}{Your API key. By default, retrieved from \code{getOption('sm_api_key')}.}
\item{oauth_token}{Your OAuth 2.0 token, as generated by \code{\link{smlogin}}. By default, retrieved from \code{getOption('sm_oauth_token')}.}
\item{...}{Other arguments passed to \code{\link[httr]{POST}}.}
}
\details{This function retrieves basic details about a user. It can be used as a hello world test after \code{\link{smlogin}}.}
\value{An object of class \code{sm_userdetails}.}
\references{
\url{https://developer.surveymonkey.com/mashery/get_user_details}
}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{
smlogin()
userdetails()
}
}
%\keyword{}
