% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_survey_by_name.R
\name{find_survey_by_name}
\alias{find_survey_by_name}
\title{find_survey_by_name}
\usage{
find_survey_by_name(name, s)
}
\arguments{
\item{name}{A string of name of the survey to be returned}

\item{s}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A single of object of class {sm_response}
}
\description{
Get responses for a SurveyMonkey survey
}
