% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.r
\name{survey_responses}
\alias{survey_responses}
\title{survey_responses}
\usage{
survey_responses(survey)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}
}
\value{
A data frame with survey responses
}
\description{
Extracts data from the survey responses data set and formats it as a data frame for analysis
}
