% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydetails.r
\name{surveydetails}
\alias{surveydetails}
\title{surveydetails}
\usage{
surveydetails(survey, question_details = TRUE,
  oauth_token = getOption("sm_oauth_token"), ...)
}
\arguments{
\item{survey}{A sm_survey object, as retrieved by \code{surveylist()}.}

\item{oauth_token}{Your OAuth 2.0 token, as generated by \code{smlogin}. By default, retrieved from \code{getOption('sm_oauth_token')}.}

\item{question_detail}{A logical value to indicate whether to include details on questions and answer choices.  By default = TRUE.}
}
\value{
A list of objects of class \code{sm_survey}.
}
\description{
Get detailed information about a survey
}
\details{
\code{sureydetails()}This function calls the SurveyMonkey API using the current oauth token and returns
details about a survey including the number of pages, questions, answer choices,
urls, etc...

\code{surveyquestions()}This function extracts a named character vector of question wordings
from a \dQuote{sm_survey} object. This can be useful for creating a codebook of responses or for mapping 
responses (from \code{\link{getresponses}}) to the original question wordings.

\code{surveypreview} opens a survey preview url in a web browser via \code{\link[utils]{browseURL}}.
}
\references{
SurveyMonkey API V3 at \url{https://developer.surveymonkey.com/api/v3/#surveys}
}
\keyword{}

