% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdetails.r
\name{userdetails}
\alias{userdetails}
\title{userdetails}
\usage{
userdetails(oauth_token = getOption("sm_oauth_token"), ...)
}
\arguments{
\item{oauth_token}{Your OAuth 2.0 token, as generated by \code{\link{smlogin}}. By default, retrieved from \code{getOption('sm_oauth_token')}.}
}
\value{
An object of class \code{sm_userdetails}.
}
\description{
Obtains information about SurveyMonkey user.
}
\details{
This function calls the SurveyMonkey API using the current oauth token and returns
information about the SurveyMonkey user and account associated with the token.  It can be used
as a "Hello World" after \code{smlogin}
}
\references{
SurveyMonkey API V3 at \url{https://developer.surveymonkey.com/api/v3/#users-me}
}

